+++ app/src/renderer/components/staking/TableValidators.vue
@@ -167,7 +167,7 @@ export default {
       address && this.updateDelegates()
     }
   },
-  async mounted() {
+  mounted() {
     Mousetrap.bind([`command+f`, `ctrl+f`], () => this.setSearch(true))
     Mousetrap.bind(`esc`, () => this.setSearch(false))
   },